backend/requirements.txt
Flask
flask-cors
openpyxl
python-dotenv
openai
xlwings